refactor(conversations): split archive state module
